Improvements to multi-processing
Apparently there have been some improvements to the debug adapter protocol specification to get some assumptions out of the way but they are not implemented in VSCode. Momentarily this is the best support I think is possible for multi-processing without needing DAP v1.51.x.  When resuming and stepping some assumptions had to be made about which thread group needed to be continued/stepped, the reason being is that it is indistinguishable if the command came from the top-ribbon or the process/threads window.